+++ components/Footer.js
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import styled from "styled-components";
-import { BookmarkIcon } from "./BookmarkButton";
+import { BookmarkIcon } from "./Icons";
 import ScheduleIcon from "./Icons";
 
 export default function Footer() {
