+++ components/Icons.js
@@ -14,3 +14,34 @@ export default function ScheduleIcon({ width, height, fill }) {
     </svg>
   );
 }
+
+export function CreateIcon({ width, height, fill }) {
+  return (
+    <svg
+      width={width}
+      height={height}
+      viewBox="0 0 20 20"
+      fill={fill}
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M8.57143 20V11.4286H0V8.57143H8.57143V0H11.4286V8.57143H20V11.4286H11.4286V20H8.57143Z"
+        fill="#245669"
+      />
+    </svg>
+  );
+}
+
+export function BookmarkIcon({ color, width, height }) {
+  return (
+    <svg
+      fill={color}
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 24 24"
+      width={width}
+      height={height}
+    >
+      <path d="M17,4v14.967l-4.212-1.805L12,16.824l-0.788,0.338L7,18.967V4H17 M17,2H7C5.9,2,5,2.9,5,4v18l7-3l7,3V4C19,2.9,18.1,2,17,2 L17,2z" />
+    </svg>
+  );
+}
